requirements.txt
fastapi
uvicorn
pandas
openpyxl
jinja2
sentence-transformers
faiss-cpu
